refactor: split types/types.go — JSON type conversion into types_json.go
